almost end of the project commit

+++ app/Http/Controllers/UserController.php
@@ -6,11 +6,12 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Store\StoreUserRequest;
 use App\Http\Requests\Update\UpdateUserRequest;
 use App\Models\Office;
-use App\Models\Role;
 use App\Models\User;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\View\View;
+use Spatie\Permission\Models\Role;
+use Spatie\Permission\Models\Permission;
 
 class UserController extends Controller
 {
@@ -21,9 +22,15 @@ class UserController extends Controller
 
     public function index()
     {
-
-        $users = User::orderBy('id','asc')->get();
+        $users = User::orderBy('id','asc')->with('roles')->get();
         return view('users.index', compact('users'));
+    }
+
+    public function show(User $user)
+    {
+        $roles = Role::all();
+        $permissions = Permission::all();
+        return view('users.role', compact('user', 'roles', 'permissions'));
     }
 
     /**
@@ -41,6 +48,44 @@ class UserController extends Controller
         return view('users.create', ['modulConf' => $modulConf], compact('offices', 'roles'));
     }
 
+    public function assignRole(Request $request, User $user)
+    {
+        if ($user->hasRole($request->role)) {
+            return back()->with('message', 'Role exists.');
+        }
+
+        $user->assignRole($request->role);
+        return back()->with('message', 'Role assigned.');
+    }
+
+    public function removeRole(User $user, Role $role)
+    {
+        if ($user->hasRole($role)) {
+            $user->removeRole($role);
+            return back()->with('message', 'Role removed.');
+        }
+
+        return back()->with('message', 'Role not exists.');
+    }
+
+    public function givePermission(Request $request, User $user)
+    {
+        if ($user->hasPermissionTo($request->permission)) {
+            return back()->with('message', 'Permission exists.');
+        }
+        $user->givePermissionTo($request->permission);
+        return back()->with('message', 'Permission added.');
+    }
+
+    public function revokePermission(User $user, Permission $permission)
+    {
+        if ($user->hasPermissionTo($permission)) {
+            $user->revokePermissionTo($permission);
+            return back()->with('message', 'Permission revoked.');
+        }
+        return back()->with('message', 'Permission does not exists.');
+    }
+
     /**
      * Store a newly created resource in storage.
      *
@@ -52,7 +97,7 @@ class UserController extends Controller
 
         User::create($request->validated());
 
-        return to_route('users.index')
+        return to_route('admin.users.index')
             ->with('toastr', [
                 'success',
                 'Yeni kullanıcı başarılı bir şekilde eklendi.',
@@ -91,7 +136,7 @@ class UserController extends Controller
     {
         $user->update($request->validated());
 
-        return to_route('users.edit', $user->id)
+        return to_route('admin.users.edit', $user->id)
             ->with('toastr', [
                 'success',
                 'Kullanıcı başarılı bir şekilde güncellendi.',
@@ -104,9 +149,13 @@ class UserController extends Controller
      * @param  User  $user
      * @return \Illuminate\Http\Response
      */
-    public function destroy(User  $user)
+    public function destroy(User $user)
     {
+        if ($user->hasRole('admin')) {
+            return back()->with('message', 'you are admin.');
+        }
         $user->delete();
-        return redirect()->route('users.index')->with('success','User has been deleted successfully');
+
+        return back()->with('message', 'User deleted.');
     }
 }
